+++ src/modules/payment.ts
@@ -13,6 +13,12 @@ type GetTransactionsParams = AutoPaginatedParams<
 type GetRefundsParams = AutoPaginatedParams<
     operations['payment_get_refunds']['parameters']['query']
 >;
+type GetLocationsParams = AutoPaginatedParams<
+    operations['payment_get_locations']['parameters']['query']
+>;
+type GetPayoutsParams = AutoPaginatedParams<
+    operations['payment_get_payouts']['parameters']['query']
+>;
 
 const paymentFactory = {
     getPayments(
@@ -67,6 +73,28 @@ const paymentFactory = {
             params,
             method: 'get',
             url: `/consumers/{consumer_id}/payment/refunds`,
+            rawData: options?.rawData,
+        };
+    },
+    getLocations(
+        params?: GetLocationsParams,
+        options?: RawDataOption
+    ): RequestData<components['schemas']['ChiftPage_PaymentLocationItem_']> {
+        return {
+            params,
+            method: 'get',
+            url: `/consumers/{consumer_id}/payment/locations`,
+            rawData: options?.rawData,
+        };
+    },
+    getPayouts(
+        params?: GetPayoutsParams,
+        options?: RawDataOption
+    ): RequestData<components['schemas']['ChiftPage_PayoutItemOut_']> {
+        return {
+            params,
+            method: 'get',
+            url: `/consumers/{consumer_id}/payment/payouts`,
             rawData: options?.rawData,
         };
     },