+++ src/modules/pos.ts
@@ -23,6 +23,8 @@ type GetAccountingCategoriesParams = AutoPaginatedParams<
 
 type GetOrdersParams = AutoPaginatedParams<operations['pos_get_orders']['parameters']['query']>;
 
+type GetTaxesParams = AutoPaginatedParams<operations['pos_get_taxes']['parameters']['query']>;
+
 const posFactory = {
     getLocations(options?: RawDataOption): RequestData<components['schemas']['POSLocationItem'][]> {
         return {
@@ -182,6 +184,17 @@ const posFactory = {
             params,
             method: 'get',
             url: `/consumers/{consumer_id}/pos/objectives`,
+            rawData: options?.rawData,
+        };
+    },
+    getTaxes(
+        params?: GetTaxesParams,
+        options?: RawDataOption
+    ): RequestData<components['schemas']['ChiftPage_POSTaxRateItem_']> {
+        return {
+            params,
+            method: 'get',
+            url: `/consumers/{consumer_id}/pos/tax-rates`,
             rawData: options?.rawData,
         };
     },